+++ lib/src/televerse/context/context.dart
@@ -62,6 +62,7 @@ part 'chat_member.dart';
 part 'poll.dart';
 part 'poll_answer.dart';
 part 'chosen_inline_result.dart';
+part 'chat_join_request.dart';
 
 /// This class is used to represent the context of an update. It contains the update and the [RawAPI] instance.
 ///