+++ lib/src/types/aliases.dart
@@ -37,3 +37,9 @@ typedef PollAnswerHandler = FutureOr<void> Function(PollAnswerContext ctx);
 /// This is used to define a callback function for almost all the methods that listen to chosen inline results.
 typedef ChosenInlineResultHandler = FutureOr<void> Function(
     ChosenInlineResultContext ctx);
+
+/// [ChatJoinRequestHandler] is a type alias for a function that takes a [ChatJoinRequestContext] as parameter and returns a [FutureOr] of void.
+///
+/// This is used to define a callback function for almost all the methods that listen to chat join requests.
+typedef ChatJoinRequestHandler = FutureOr<void> Function(
+    ChatJoinRequestContext ctx);